Relocate mislocated test classes and ratchet the mockless share
An audit prompted by the mandate slice undercount found fifteen test
files living away from their subjects: the four stale mandate-side
signature specs fall to the superseding signature-module suite, the
savings root, holdings, role, and fund tests move beside their classes,
and the TulevaFundTest variants merge. The ledger validator tests
deliberately stay in the ledger root where their package-private
builders live. The mock-free ratchet becomes a share rather than a
count so deleting redundant tests is never penalized.
