+++ metrics/scorecard.json
@@ -9,11 +9,12 @@
   "disconnectedClasses": 454,
   "fqnViolations": 0,
   "godClasses": 10,
-  "interactionVerifyingTestClasses": 360,
+  "interactionVerifyingTestClasses": 357,
   "lineCoverage": 83.77,
   "longClasses": 3,
   "meanInstability": 0.473,
-  "mocklessTestClasses": 452,
+  "mocklessTestClasses": 450,
+  "mocklessTestShare": 46.1,
   "moduleCycles": 0,
   "moduleSliceTests": 0,
   "modulithViolations": 0,
@@ -21,7 +22,7 @@
   "mutationScoreSliced": 88.55,
   "mutationsSliced": 13028,
   "mutationsTotal": 13028,
-  "overMockedTestClasses": 111,
+  "overMockedTestClasses": 107,
   "pitestSlicedModules": 52,
   "pmdViolations": 144,
   "pmdViolationsByRule": {
@@ -31,14 +32,14 @@
     "GodClass": 10,
     "TooManyMethods": 4
   },
-  "slowest10TestClassesSeconds": 111.9,
+  "slowest10TestClassesSeconds": 111.3,
   "springBootTests": 76,
-  "suiteWallTimeSeconds": 171,
+  "suiteWallTimeSeconds": 188,
   "testClassesOver10Seconds": 5,
-  "testClassesTotal": 982,
-  "testIntegrationShare": 6.8,
+  "testClassesTotal": 976,
+  "testIntegrationShare": 6.9,
   "testSliceShare": 14.1,
-  "testUnitShare": 79.1,
+  "testUnitShare": 79.0,
   "top15ClassLines": 6875,
   "totalPackages": 227,
   "unmarkedPackages": 0
