--- src/test/groovy/ee/tuleva/onboarding/error/response/InputErrorsConverterSpec.groovy
@@ -1,5 +1,0 @@
-package ee.tuleva.onboarding.error.response
-
-class TestCommand {
-    String someField
-}
