+++ src/test/groovy/ee/tuleva/onboarding/savings/SavingFundDeadlinesServiceSpec.groovy
@@ -1,4 +1,4 @@
-package ee.tuleva.onboarding.savings.fund
+package ee.tuleva.onboarding.savings
 
 
 import ee.tuleva.onboarding.savings.SavingFundDeadlinesService
