+++ src/test/groovy/ee/tuleva/onboarding/savings/SavingsFundFeesTest.java
@@ -1,12 +1,10 @@
-package ee.tuleva.onboarding.savings.fund;
+package ee.tuleva.onboarding.savings;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.BDDMockito.given;
 
 import ee.tuleva.onboarding.fund.Fund;
 import ee.tuleva.onboarding.fund.FundRepository;
-import ee.tuleva.onboarding.savings.SavingsFundConfiguration;
-import ee.tuleva.onboarding.savings.SavingsFundFees;
 import java.math.BigDecimal;
 import java.util.Locale;
 import org.junit.jupiter.api.BeforeEach;
